[refs #00000] update html color, fix js/html, move font
The html body is all one color, a shade of black `#191A1B`
The time now shows in event listings. Still need to fix format.
Moving the font file to the styles folder fixes the error that kept
popping up
Also added the result id to the html so that the events actually show up
on the web page.